Move builder logic to enum
