+++ graphqlapi/src/main/java/io/stargate/graphql/schema/fetchers/dml/DmlFetcher.java
@@ -9,15 +9,11 @@ import io.stargate.auth.AuthenticationService;
 import io.stargate.db.Persistence;
 import io.stargate.db.schema.Column;
 import io.stargate.db.schema.Table;
-import io.stargate.graphql.schema.FilterOperator;
 import io.stargate.graphql.schema.NameMapping;
 import io.stargate.graphql.schema.fetchers.CassandraFetcher;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public abstract class DmlFetcher<ResultT> extends CassandraFetcher<ResultT> {
 
@@ -45,28 +41,10 @@ public abstract class DmlFetcher<ResultT> extends CassandraFetcher<ResultT> {
 
       for (Map.Entry<String, Object> conditionMap : clauseEntry.getValue().entrySet()) {
         FilterOperator operator = FilterOperator.fromFieldName(conditionMap.getKey());
-        Condition condition;
-        if (operator != FilterOperator.IN) {
-          condition =
-              operator.buildCondition(column.name(), toCqlTerm(column, conditionMap.getValue()));
-        } else {
-          condition =
-              operator.buildCondition(
-                  column.name(), buildListLiterals(column, conditionMap.getValue()));
-        }
-        clause.add(condition);
+        clause.add(operator.buildCondition(column, conditionMap.getValue(), nameMapping));
       }
     }
     return clause;
-  }
-
-  private List<Term> buildListLiterals(Column column, Object o) {
-    if (o instanceof Collection<?>) {
-      Collection<?> values = (Collection<?>) o;
-      return values.stream().map(item -> toCqlTerm(column, item)).collect(Collectors.toList());
-    }
-
-    return Collections.singletonList(toCqlTerm(column, o));
   }
 
   protected List<Relation> buildFilterConditions(
@@ -78,31 +56,8 @@ public abstract class DmlFetcher<ResultT> extends CassandraFetcher<ResultT> {
     for (Map.Entry<String, Map<String, Object>> clauseEntry : columnList.entrySet()) {
       Column column = getColumn(table, clauseEntry.getKey());
       for (Map.Entry<String, Object> condition : clauseEntry.getValue().entrySet()) {
-        Relation relation;
         FilterOperator operator = FilterOperator.fromFieldName(condition.getKey());
-        if (operator == FilterOperator.IN) {
-          relation =
-              operator.buildRelation(
-                  column.name(), buildListLiterals(column, condition.getValue()));
-        } else if (operator == FilterOperator.CONTAINS) {
-          relation =
-              operator.buildRelation(column.name(), toCqlElementTerm(column, condition.getValue()));
-        } else if (condition.getKey().equals("containsKey")) {
-          relation =
-              operator.buildRelation(column.name(), toCqlKeyTerm(column, condition.getValue()));
-        } else if (condition.getKey().equals("containsEntry")) {
-          Column.ColumnType mapType = column.type();
-          assert mapType != null && mapType.isMap();
-          Map<String, Object> entry = (Map<String, Object>) condition.getValue();
-          Column.ColumnType keyType = mapType.parameters().get(0);
-          Term keyTerm = toCqlTerm(keyType, entry.get("key"));
-          Column.ColumnType valueType = mapType.parameters().get(1);
-          Term valueTerm = toCqlTerm(valueType, entry.get("value"));
-          relation = operator.buildRelation(column.name(), keyTerm, valueTerm);
-        } else {
-          relation = operator.buildRelation(column.name(), toCqlTerm(column, condition.getValue()));
-        }
-        relations.add(relation);
+        relations.add(operator.buildRelation(column, condition.getValue(), nameMapping));
       }
     }
     return relations;
@@ -140,24 +95,6 @@ public abstract class DmlFetcher<ResultT> extends CassandraFetcher<ResultT> {
   }
 
   protected Term toCqlTerm(Column column, Object value) {
-    return toCqlTerm(column.type(), value);
-  }
-
-  private Term toCqlElementTerm(Column column, Object value) {
-    Column.ColumnType collectionType = column.type();
-    assert collectionType != null && collectionType.isCollection();
-    Column.ColumnType elementType = collectionType.parameters().get(collectionType.isMap() ? 1 : 0);
-    return toCqlTerm(elementType, value);
-  }
-
-  private Term toCqlKeyTerm(Column column, Object value) {
-    Column.ColumnType mapType = column.type();
-    assert mapType != null && mapType.isMap();
-    Column.ColumnType keyType = mapType.parameters().get(0);
-    return toCqlTerm(keyType, value);
-  }
-
-  private Term toCqlTerm(Column.ColumnType type, Object value) {
-    return DataTypeMapping.toCqlTerm(type, value, nameMapping);
+    return DataTypeMapping.toCqlTerm(column.type(), value, nameMapping);
   }
 }
